deployment/requirements-heroku.txt
Flask==2.3.3
Flask-Session==0.5.0
Werkzeug==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
python-dotenv==1.0.0
google-generativeai==0.3.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
